# service/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  resendMail.ResendMailRequest:
    properties:
      email:
        example: email@mail.com
        type: string
      firstName:
        example: สมชาย
        type: string
    required:
    - email
    - firstName
    type: object
  standardDevelopController.AddStandardDevelopRequest:
    properties:
      ageRange:
        example: 24
        type: integer
      developments:
        example:
        - '[{"category":"กล้ามเนื้อ"'
        - '"detail":"วิ่งได้"'
        - '"image":"https://example.com/crawl.png"'
        - '"note":"ควรฝึกทุกวัน"}]'
        items:
          type: string
        type: array
    required:
    - ageRange
    - developments
    type: object
  standardDevelopController.UpdateStandardDevelopRequest:
    properties:
      ageRange:
        example: 24
        type: integer
      developments:
        example:
        - '[{"category":"กล้ามเนื้อ"'
        - '"detail":"วิ่งได้"'
        - '"image":"https://example.com/crawl.png"'
        - '"note":"ควรฝึกทุกวัน"}]'
        items:
          type: string
        type: array
    required:
    - developments
    type: object
  standardDevelopModel.DevelopmentItem:
    properties:
      category:
        type: string
      detail:
        type: string
      image:
        type: string
      note:
        type: string
    type: object
  standardDevelopModel.StandardDevelop:
    properties:
      ageRange:
        type: integer
      developments:
        items:
          $ref: '#/definitions/standardDevelopModel.DevelopmentItem'
        type: array
      id:
        type: string
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: Backend API for MimiCare Project (Gin + MongoDB)
  title: MimiCare API
  version: "1.0"
paths:
...
  /resendMail/send-password:
    post:
      consumes:
      - application/json
      description: ใช้สำหรับส่งรหัสผ่านใหม่ไปยังอีเมลของผู้ใช้ (กรณีลืมรหัสผ่าน)
      parameters:
      - description: ข้อมูลอีเมลและชื่อผู้ใช้
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/resendMail.ResendMailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: ส่งรหัสผ่านใหม่ไปยังอีเมลของผู้ใช้
      tags:
      - ResendMail
  /standardDevelop/add:
    post:
      consumes:
      - application/json
      description: ใช้สำหรับเพิ่มข้อมูลพัฒนาการมาตรฐานใหม่ (admin เท่านั้น)
      parameters:
      - description: ข้อมูลพัฒนาการมาตรฐาน
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/standardDevelopController.AddStandardDevelopRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: เพิ่มข้อมูลพัฒนาการมาตรฐานใหม่
      tags:
      - StandardDevelop
  /standardDevelop/all:
    get:
      description: ใช้สำหรับดึงข้อมูลพัฒนาการมาตรฐานทั้งหมด (admin เท่านั้น)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: ดึงข้อมูลพัฒนาการมาตรฐานทั้งหมด
      tags:
      - StandardDevelop
  /standardDevelop/delete/{id}:
    delete:
      description: ใช้สำหรับลบข้อมูลพัฒนาการมาตรฐานตาม ID (admin เท่านั้น)
      parameters:
      - description: ObjectID ของพัฒนาการมาตรฐาน
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: ลบข้อมูลพัฒนาการมาตรฐานตาม ID
      tags:
      - StandardDevelop
  /standardDevelop/get/{id}:
    get:
      description: ใช้สำหรับดึงข้อมูลพัฒนาการมาตรฐานตาม ID (admin เท่านั้น)
      parameters:
      - description: ObjectID ของพัฒนาการมาตรฐาน
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/standardDevelopModel.StandardDevelop'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: ดึงข้อมูลพัฒนาการมาตรฐานตาม ID
      tags:
      - StandardDevelop
  /standardDevelop/update/{id}:
    put:
      consumes:
      - application/json
      description: ใช้สำหรับอัปเดตข้อมูลพัฒนาการมาตรฐานตาม ID (admin เท่านั้น)
      parameters:
      - description: ObjectID ของพัฒนาการมาตรฐาน
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลใหม่
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/standardDevelopController.UpdateStandardDevelopRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: อัปเดตข้อมูลพัฒนาการมาตรฐานตาม ID
      tags:
      - StandardDevelop
...
schemes:
- http
swagger: "2.0"